requirements.txt
# Web-Framework
flask==3.0.2
requests==2.31.0

# Datenbank
sqlalchemy==2.0.28
psycopg2-binary==2.9.9

# Datenverarbeitung
pandas==2.2.1
numpy==1.26.4

# Testing (zusätzlich zu pytest im Dockerfile)
pytest-mock==3.12.0
pytest-asyncio==0.23.5

# API-Tools
httpx==0.27.0
pydantic==2.6.3

# Utilities
python-dotenv==1.0.1
click==8.1.7

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0